+++ next.config.ts
@@ -1,7 +1,14 @@
 import type { NextConfig } from "next";
+import path from "path";
 
 const nextConfig: NextConfig = {
   /* config options here */
+  webpack: (config) => {
+    config.resolve = config.resolve ?? {};
+    config.resolve.alias = config.resolve.alias ?? {};
+    config.resolve.alias["canvas"] = path.resolve(__dirname, "./empty-module.js");
+    return config;
+  },
 };
 
 export default nextConfig;